removed db connection from global scope and made it local to function w/in router
